backend/jobfinder/requirements.txt
Django==4.0.1
python-dotenv==0.19.0
django-filter==2.4.0
djangorestframework==3.12.4
drf-yasg==1.20.0
psycopg2-binary==2.9.1
pytz==2021.3
djangorestframework-simplejwt==4.8.0